Added the drawer and customized the icon for the drawer. The drawer now uses the African logo instead of the default drawer icon.
Added the appbar also. I tried creating a method for appbar but i encountered a problem when calling it in the main page.
I was supposed to use a 'PreferredSizeWidget' instead od 'Widget' because it needs to know the preferred height of the app bar to correctly display it.
i used the default height though, looks good that way.